move all util functions to utils.py
